blockgroesse auf INT
git-svn-id: https://randi2.svn.sourceforge.net/svnroot/randi2/randi/trunk/RANDI2@1847 ccfe945e-4522-0410-b90e-900a8cb55c58

+++ db/create_db.sql
@@ -84,7 +84,7 @@ CREATE TABLE Studie (
   enddatum DATE NOT NULL,
   studienprotokoll VARCHAR(255) NOT NULL,
   status_Studie VARCHAR(25) NOT NULL,
-  blockgroesse SMALLINT NULL,
+  blockgroesse INT NULL,
   PRIMARY KEY(studienID),
   INDEX Studie_FKIndex1(Benutzerkonto_benutzerkontenID),
   FOREIGN KEY(Benutzerkonto_benutzerkontenID)
